requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
mermaid-cli>=0.1.0
pydantic>=1.10.7
